Added next steps
As of now this is a good work sample of my process to answer the most common data question I see:  "How does the price of this listing vary with the value of it's <property>?  In this case, I've answered the question "How does the price of an 18650 battery vary with the length of it's listing title?

It's been a fun few hours, and I'm in a good place to let this sit on the shelf while I work on more important projects.  If I were in an office, I'd report an answer to my supervisor's question as of now, attaching the simple scatterplot.  If it were being presented externally as a finding, I'd want to replot with seaborn to include a regression and CI band.  By adding some minor webscraping features, I could answer the "$/mAh" question, and pick the most cost effective battery to purchase w/r/t:power bank capacity.